move U*Gate to gate_one_qubit
