+++ qulacs/gate/gate_matrix.hpp
@@ -11,65 +11,6 @@
 
 namespace qulacs {
 namespace internal {
-class U1GateImpl : public OneQubitGateBase {
-    double _lambda;
-    matrix_2_2 _matrix;
-
-public:
-    U1GateImpl(UINT target, double lambda) : OneQubitGateBase(target), _lambda(lambda) {
-        _matrix = get_IBMQ_matrix(0, 0, lambda);
-    };
-
-    double lambda() const { return _lambda; }
-
-    Gate copy() const override { return std::make_shared<U1GateImpl>(*this); }
-    Gate get_inverse() const override { return std::make_shared<U1GateImpl>(_target, -_lambda); }
-
-    void update_quantum_state(StateVector& state_vector) const override;
-};
-class U2GateImpl : public OneQubitGateBase {
-    double _phi, _lambda;
-    matrix_2_2 _matrix;
-
-public:
-    U2GateImpl(UINT target, double phi, double lambda)
-        : OneQubitGateBase(target), _phi(phi), _lambda(lambda) {
-        _matrix = get_IBMQ_matrix(PI() / 2.0, phi, lambda);
-    };
-
-    double phi() const { return _phi; }
-    double lambda() const { return _lambda; }
-
-    Gate copy() const override { return std::make_shared<U2GateImpl>(*this); }
-    Gate get_inverse() const override {
-        return std::make_shared<U2GateImpl>(_target, -_lambda - PI(), -_phi + PI());
-    }
-
-    void update_quantum_state(StateVector& state_vector) const override;
-};
-
-class U3GateImpl : public OneQubitGateBase {
-    double _theta, _phi, _lambda;
-    matrix_2_2 _matrix;
-
-public:
-    U3GateImpl(UINT target, double theta, double phi, double lambda)
-        : OneQubitGateBase(target), _theta(theta), _phi(phi), _lambda(lambda) {
-        _matrix = get_IBMQ_matrix(theta, phi, lambda);
-    };
-
-    double theta() const { return _theta; }
-    double phi() const { return _phi; }
-    double lambda() const { return _lambda; }
-
-    Gate copy() const override { return std::make_shared<U3GateImpl>(*this); }
-    Gate get_inverse() const override {
-        return std::make_shared<U3GateImpl>(_target, -_theta, -_lambda, -_phi);
-    }
-
-    void update_quantum_state(StateVector& state_vector) const override;
-};
-
 class OneQubitMatrixGateImpl : public OneQubitGateBase {
     matrix_2_2 _matrix;
 
@@ -139,9 +80,6 @@ public:
 };
 }  // namespace internal
 
-using U1Gate = internal::GatePtr<internal::U1GateImpl>;
-using U2Gate = internal::GatePtr<internal::U2GateImpl>;
-using U3Gate = internal::GatePtr<internal::U3GateImpl>;
 using OneQubitMatrixGate = internal::GatePtr<internal::OneQubitMatrixGateImpl>;
 using TwoQubitMatrixGate = internal::GatePtr<internal::TwoQubitMatrixGateImpl>;
 }  // namespace qulacs
