+++ qulacs/gate/update_ops.hpp
@@ -68,7 +68,11 @@ void double_qubit_dense_matrix_gate(UINT target0,
                                     const matrix_4_4& matrix,
                                     StateVector& state);
 
-void u_gate(UINT target_qubit_index, const matrix_2_2& matrix, StateVector& state);
+void u1_gate(UINT target_qubit_index, double lambda, StateVector& state);
+
+void u2_gate(UINT target_qubit_index, double phi, double lambda, StateVector& state);
+
+void u3_gate(UINT target_qubit_index, double theta, double phi, double lambda, StateVector& state);
 
 void swap_gate(UINT target1, UINT target2, StateVector& state);
