+++ qulacs/gate/update_ops_quantum_matrix.cpp
@@ -15,8 +15,17 @@ matrix_2_2 get_IBMQ_matrix(double theta, double phi, double lambda) {
     return {cos_val, -exp_val2 * sin_val, exp_val1 * sin_val, exp_val1 * exp_val2 * cos_val};
 }
 
-void u_gate(UINT target_qubit_index, const matrix_2_2& matrix, StateVector& state) {
-    single_qubit_dense_matrix_gate(target_qubit_index, matrix, state);
+void u1_gate(UINT target_qubit_index, double lambda, StateVector& state) {
+    single_qubit_dense_matrix_gate(target_qubit_index, get_IBMQ_matrix(0., 0., lambda), state);
+}
+
+void u2_gate(UINT target_qubit_index, double phi, double lambda, StateVector& state) {
+    single_qubit_dense_matrix_gate(
+        target_qubit_index, get_IBMQ_matrix(PI() / 2., phi, lambda), state);
+}
+
+void u3_gate(UINT target_qubit_index, double theta, double phi, double lambda, StateVector& state) {
+    single_qubit_dense_matrix_gate(target_qubit_index, get_IBMQ_matrix(theta, phi, lambda), state);
 }
 }  // namespace internal
 }  // namespace qulacs
